Finally finished adding meshes

+++ include/nova_renderer/renderables.hpp
@@ -34,7 +34,7 @@ namespace nova::renderer {
         std::vector<uint32_t> indices;
     };
 
-    using MeshId = uint32_t;
+    using MeshId = uint64_t;
 
     struct StaticMeshRenderableUpdateData {
         std::string material_name;
